feat(*): 북마크 리스트 조회 로직 위치 변경
